# ai-project/api-spec.yml
openapi: 3.0.3
info:
  title: AI Chains Diagram API
  description: REST API для управления диаграммами с узлами и связями
  version: 1.0.0
  contact:
    name: AI Chains Team
    email: support@aichains.com

servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.aichains.com/v1
    description: Production server

paths:
  /templates:
    get:
      summary: Получить список всех диаграмм
      description: Возвращает список всех диаграмм на сервере
      operationId: getTemplates
      tags:
        - Templates
      responses:
        "200":
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiagramListItem"
              example:
                - id: 1
                  name: "Диаграмма анализа данных"
                - id: 2
                  name: "Диаграмма обработки текста"
        "400":
          description: Ошибка запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /templates/create:
    post:
      summary: Создать новую диаграмму
      description: Создает новую пустую диаграмму
      operationId: createTemplate
      tags:
        - Templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Название новой диаграммы
                  example: "Новая диаграмма"
                description:
                  type: string
                  description: Описание диаграммы
                  example: "Описание новой диаграммы"
      responses:
        "201":
          description: Диаграмма успешно создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    description: ID созданной диаграммы
                    example: 3
                required:
                  - id
        "400":
          description: Ошибка запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /templates/{id}/save:
    post:
      summary: Сохранить диаграмму
      description: Сохраняет изменения в существующей диаграмме
      operationId: saveTemplate
      tags:
        - Templates
      parameters:
        - name: id
          in: path
          required: true
          description: ID диаграммы
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Diagram"
            example:
              id: 1
              name: "Обновленная диаграмма"
              description: "Описание обновленной диаграммы"
              nodes:
                - id: 1
                  id_ai: 100
                  name: "Входной узел"
                  description: "Узел для ввода данных"
                  text_in: "Введите данные"
                  text_out: "Данные получены"
              lines:
                - id: 1
                  id_node_begin: 1
                  id_node_end: 2
                  description: "Связь обработки"
      responses:
        "200":
          description: Диаграмма успешно сохранена
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    description: ID диаграммы
                    example: 1
                  name:
                    type: string
                    description: Название диаграммы
                    example: "Сохраненная диаграмма"
                required:
                  - id
                  - name
        "400":
          description: Ошибка запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Диаграмма не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /templates/{id}/delete:
    delete:
      summary: Удалить диаграмму
      description: Удаляет диаграмму по указанному ID
      operationId: deleteTemplate
      tags:
        - Templates
      parameters:
        - name: id
          in: path
          required: true
          description: ID диаграммы
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "204":
          description: Диаграмма успешно удалена
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    description: ID удаленной диаграммы
                    example: 1
                  name:
                    type: string
                    description: Название удаленной диаграммы
                    example: "Удаленная диаграмма"
                required:
                  - id
                  - name
        "400":
          description: Ошибка запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Диаграмма не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Node:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID узла
          example: 1
        id_ai:
          type: integer
          format: int64
          description: Связанная AI-сущность
          example: 100
        name:
          type: string
          description: Короткое отображаемое имя узла
          example: "Входной узел"
        description:
          type: string
          description: Описание узла (всплывающее описание и т.п.)
          example: "Узел для ввода данных"
        text_in:
          type: string
          description: Входной текст
          example: "Введите данные для анализа"
        text_out:
          type: string
          description: Выходной текст
          example: "Данные обработаны"
      required:
        - id
        - id_ai
        - name
        - description
        - text_in
        - text_out

    Line:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID линии
          example: 1
        id_node_begin:
          type: integer
          format: int64
          description: ID начального узла
          example: 1
        id_node_end:
          type: integer
          format: int64
          description: ID конечного узла
          example: 2
        description:
          type: string
          description: Описание линии
          example: "Связь обработки"
      required:
        - id
        - id_node_begin
        - id_node_end
        - description

    Diagram:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID диаграммы
          example: 1
        name:
          type: string
          description: Короткое отображаемое имя диаграммы
          example: "Диаграмма анализа данных"
        description:
          type: string
          description: Описание диаграммы (для чего применяется и т.п.)
          example: "Диаграмма для анализа текстовых данных"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
          description: Массив узлов
        lines:
          type: array
          items:
            $ref: "#/components/schemas/Line"
          description: Массив линий
      required:
        - id
        - name
        - description
        - nodes
        - lines

    DiagramListItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID диаграммы
          example: 1
        name:
          type: string
          description: Короткое отображаемое имя диаграммы
          example: "Диаграмма анализа данных"
      required:
        - id
        - name

    Error:
      type: object
      properties:
        code:
          type: integer
          description: Код ошибки
          example: 400
        message:
          type: string
          description: Короткое сообщение
          example: "Неверный запрос"
        details:
          type: string
          description: Детали ошибки
          example: "Поле 'name' обязательно для заполнения"
      required:
        - code
        - message
        - details

tags:
  - name: Templates
    description: Операции с диаграммами